fix: don't crash when template has no frontmatter
the yaml_front_matter crate says frontmatter is optional but in practice
expects it to always be there. no specific error values are raised so
it's difficult to differentiate this error from invalid YAML.
for now this workaround will allow whiskers to continue when no
frontmatter is present.
long term, a better solution would be to contribute to or fork
yaml_front_matter to make `metadata` an `Option<T>` instead of `T`